weather-reporter/requirements.txt
flask
python-dotenv
requests
concurrent-future